requirements.txt
aiogram==3.17.0
requests==2.32.3
python-dotenv==1.0.1
matplotlib>=3.5.0
